+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - listContent() only returned writable modules
 - refreshMTimes() hangs on unavailable modules
+- Better tracking of external changes using dft #245
+
+### Changed
 - Update specification versions
+
 
 ## [0.8.0] - 2020-09-08
 ### Fixed